Enemy HP bar now working
The HP bar had a bug that caused to flicker. This was fixed by setting the HP bar to check position every fixed update instead of update. The enemy attack function is what rotates the enemy, and that is called every fixed update. The desync is what caused the flickering.