Pause Menu fully working
Hitting escape on the keyboard pauses the game and opens the pause menu. On the pause menu there is a button to resume, which pauses the game and closes the pause menu. The pause menu can also return to main menu. There was a problem in which SnowPrincess would still spawn snowballs even when the game was paused but that has been fixed by checking if the game is paused before calling the attack function