Detached the HUD from SnowPrincess and added Pause Menu to it
This reorganization is so all HUD overlays and menus go on one prefab to avoid the need to create multiple canvases for each new level. It also puts all HUD related code in one location, and makes it so that if SnowPrincess or any  other object is removed from the game, the HUD and all related UI elements continue to work.