Created damage field
As of now, the damage field acts like a giant snowball, next change is to make it placeable from the build menu. After that, make it so the damage is delayed by a second, and then after that make is so the object is destroyed after the delay even if it hasn't damaged anything. Much later I'll have an animation for it.